+++ WeChatFerry/spy/send_msg.cpp
@@ -10,7 +10,7 @@
 
 extern HANDLE g_hEvent;
 extern WxCalls_t g_WxCalls;
-extern DWORD g_WeChatWinDllAddr;
+extern UINT64 g_WeChatWinDllAddr;
 extern string GetSelfWxid(); // Defined in spy.cpp
 #if 0
 void SendTextMessage(string wxid, string msg, string atWxids)
